ALTUSGROUPSUB-44: Add Content-Type to TokenAPI
Fixes https://support.wso2.com/jira/browse/ALTUSGROUPSUB-44
Fixes https://github.com/wso2/product-apim/issues/3773

Handler property is introduced to decide if we need to add the
Content-Type header to the request. This is because client is
expected to send the Content-Type header in the request and we
should only manipulate the header when requests comes in from
a poorly implemented client.