Fix API import when publisher access control enabled
When importing multiple versions of APIs with publisher
access control enabled by a different user than the API
provider, role validation is done based on the provider.
Previously provider was taken from the artefact which
was the original user and if that user does not exist in the
new environment, there were multiple errors and this
sets the provider to the logged in user.

Resolves: https://github.com/wso2/api-manager/issues/1020